+++ prometheus/testutil/testutil_test.go
@@ -1,4 +1,4 @@
-package testutils
+package testutil
 
 import (
 	"testing"
